#39 Move serializer and repository (now arxiv_parser) to interface_adapter and fix imports
