+++ tests/interface_adapters/serializers/test_arxiv_document_serializer.py
@@ -2,7 +2,7 @@ import datetime
 import json
 import pytest
 
-from webminer.serializers.json import arxiv_document_serializer as srs
+from webminer.interface_adapters.serializers.json import arxiv_document_serializer as srs
 from webminer.entities.arxiv_document import ArxivDocument
 
 test_url = "https://arxiv.org/abs/1801.06605"
