+++ webminer/interface_adapters/serializers/json/arxiv_document_serializer.py
